dev: fix max ethernet size issue for VFs
When user provides maximum ethernet size more than supported
size then initially it gets stored in a u32 variable which is
further copied to u16 variable. Due to this conversion
maximum input value truncated to 0. This trunctation leads to
failure in later stage.

Type: fix
JIRA: https://essjira.marvell.com/browse/IPBUSW-62958

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/toolkits/vpp/+/147519
Tested-by: sa_ip-toolkits-Jenkins <sa_ip-toolkits-jenkins@marvell.com>
(cherry picked from commit aded3e60710c9fec40925cbee9cb123fc7f5dd2e)
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/toolkits/vpp/+/147615
Tested-by: Monendra Singh Kushwaha <kmonendra@marvell.com>

+++ src/vnet/dev/dev.h
@@ -170,7 +170,7 @@ typedef struct vnet_dev_port_cfg_change_req
   {
     u8 promisc : 1;
     vnet_dev_hw_addr_t addr;
-    u16 max_rx_frame_size;
+    u32 max_rx_frame_size;
     vnet_dev_queue_id_t queue_id;
     struct
     {
